Add a mu translation module
