+++ src/command_line/procedures.rs
@@ -11,7 +11,9 @@ use {
         convenience::{apply::Apply, compose::Compose},
         simplifying::fol::{classic::CLASSIC, ht::HT, intuitionistic::INTUITIONISTIC},
         syntax_tree::{asp, fol, Node as _},
-        translating::{completion::completion, gamma::gamma, natural::mu, tau_star::tau_star},
+        translating::{
+            completion::completion, gamma::gamma, mu::mu, natural::natural, tau_star::tau_star,
+        },
         verifying::{
             prover::{vampire::Vampire, Prover, Report, Status, Success},
             task::{
@@ -87,16 +89,24 @@ pub fn main() -> Result<()> {
                     print!("{gamma_theory}")
                 }
 
-                Translation::TauStar => {
-                    let program =
-                        input.map_or_else(asp::Program::from_stdin, asp::Program::from_file)?;
-                    let theory = tau_star(program);
-                    print!("{theory}")
-                }
                 Translation::Mu => {
                     let program =
                         input.map_or_else(asp::Program::from_stdin, asp::Program::from_file)?;
                     let theory = mu(program);
+                    print!("{theory}")
+                }
+
+                Translation::Natural => {
+                    let program =
+                        input.map_or_else(asp::Program::from_stdin, asp::Program::from_file)?;
+                    let theory = natural(program).context("the given program is not regular")?;
+                    print!("{theory}")
+                }
+
+                Translation::TauStar => {
+                    let program =
+                        input.map_or_else(asp::Program::from_stdin, asp::Program::from_file)?;
+                    let theory = tau_star(program);
                     print!("{theory}")
                 }
             }
