+++ src/translating/mod.rs
@@ -1,4 +1,5 @@
 pub mod completion;
 pub mod gamma;
+pub mod mu;
 pub mod natural;
 pub mod tau_star;
